Updating code to run more efficiently
I got rid of additional fluff in the code that saves processing power to reduce GPU usage and produce faster run time as this program should only focus on completing woodcutting as its task. Added a random aspect to some of the processes to increase the productivity of the program instead of waiting 30+ seconds on tasks.