Added a bell sound to notify user
This bell sound is used to notify the user that the inventory is full of willow logs and is ready to be set back to bank either manually or with the program taking the reins. Since it would be easier for the user to move the in game character to the bank rather than the wait time for the program to do it and potentially mess up as it sometimes messed up during our runs of the program. It would save a few seconds every time the program runs rather than waiting around doing nothing as the program tries to find its way to bank.